pyproject.toml
[project]
name = "meshtastic-zigbee"
version = "0.1.0"
description = "MQTT to Meshtastic bridge for Zigbee2MQTT events"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "paho-mqtt>=1.6.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
